refactor(oneseo): EvaluationStatus enum class 위치를 oneseo.entity.type 패키지로 변경
